requirements-dev.txt
# Code quality and testing
black==23.12.1
flake8==7.0.0
mypy==1.8.0
isort==5.13.2
PyQt5-stubs==5.15.6.0
types-setuptools==75.2.0.20241025

# Testing
pytest==8.3.3
pytest-cov==5.0.0
pytest-html==4.1.1
pytest-metadata==3.1.1
pytest-mock==3.14.0
pytest-xdist==3.6.1
coverage==7.6.1

# Build and distribution
build==1.0.3
twine==4.0.2
setuptools==75.1.0
